requirements.txt
-r requirements/conda.txt

click
path.py
pysam>=0.7.5
toml
toolz
